requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Amazon SP API
python-amazon-sp-api==1.9.5

# AI response generation
anthropic==0.42.0

# HTTP client
httpx==0.28.1

# Settings & environment
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Background tasks
apscheduler==3.10.4

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
